+++ packages/harch-web/src/HarchWeb/Api.hs
@@ -44,6 +44,7 @@ module HarchWeb.Api
     parseApiField,
     queryField,
     headerField,
+    cookieField,
     requiredField,
     optionalField,
     fieldWithDefault,